requirements.txt
Flask==3.1.0
Flask-WTF==1.2.2
WTForms==3.2.1
email_validator==2.2.0
Flask-MySQLdb==2.0.0
mysqlclient==2.2.7
Bootstrap-Flask==2.4.1
Jinja2==3.1.5
MarkupSafe==3.0.2
Werkzeug==3.1.3
click==8.1.8
colorama==0.4.6
blinker==1.9.0
greenlet==3.1.1
dominate==2.9.1
dnspython==2.7.0
numpy==2.2.2
typing_extensions==4.12.2
visitor==0.1.3
itsdangerous==2.2.0
